He mejorado el aplicativo controlando los errores de asociación entre entidades para que no pete el aplicativo. He añadido DTOs para enviar JSON datos para las busquedas con AJAX en JS

+++ src/main/java/com/smarvel/springboot/backend/serviceImpl/ProductoServiceImpl.java
@@ -50,6 +50,11 @@ public class ProductoServiceImpl implements IProductoService{
 	@Override
 	@Transactional
 	public void borrarProducto(Long id) {
+		boolean tieneSupermercado = productoSupermercadoRepository.existsByProductoId(id);
+	    
+	    if(tieneSupermercado) {
+	        throw new IllegalStateException("No se puede borrar este producto porque está asociado a un supermercado.");
+	    }
 		productoRepository.deleteById(id);
 		
 	}
